Fixed an obvious problem with update.
The way it was before, a shopping cart would try to update itself even
when it was also supposed to run the script, which is a waste of effort.
It also tried to create a connection and update the cart even before the
script was run, which is stupid. Should be fixed now.